Update potentially sexist language
A friend of mine pointed out that the original phrasing could be
interpreted as women being incapable of operating the Voder. My intent
was to elevate and bolster the few people who were capable of operating
the Voder, which was exclusively women. The Voder was seen as the future
of telephony so therefore a natural progression of the phone operator
which was seen as a woman's job back in the 1930s.

Discussing historical sexism is difficult and way more nuanced than I
realized when I wrote this up a few years ago, so I really appreciate my
friend's input here and I'm glad I can fix it and move on.